Add support to detect remote & origin automatically
